Update UI : Improved package structure
