+++ composeApp/src/commonMain/kotlin/com/yong/taximeter/ui/main/subscreen/home/HomeViewModel.kt
@@ -1,7 +1,7 @@
-package com.yong.taximeter.ui.main.subpage.store
+package com.yong.taximeter.ui.main.subscreen.home
 
 import cafe.adriel.voyager.core.model.ScreenModel
 
-class StoreViewModel: ScreenModel {
+class HomeViewModel: ScreenModel {
     // TODO: ViewModel 로직 구현
 }
