+++ composeApp/src/commonMain/kotlin/com/yong/taximeter/ui/main/subscreen/setting/SettingScreen.kt
@@ -1,4 +1,4 @@
-package com.yong.taximeter.ui.main.subpage.home
+package com.yong.taximeter.ui.main.subscreen.setting
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -7,15 +7,15 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import cafe.adriel.voyager.core.screen.Screen
 
-object HomeScreen: Screen {
+object SettingScreen: Screen{
     @Composable
     override fun Content() {
         Box(
             modifier = Modifier
                 .fillMaxSize()
         ) {
-            // TODO: Home UI 구현
-            Text("Home Screen")
+            // TODO: Setting UI 구현
+            Text("Setting Screen")
         }
     }
 }
