+++ composeApp/src/commonMain/kotlin/com/yong/taximeter/ui/main/subscreen/store/StoreViewModel.kt
@@ -1,7 +1,7 @@
-package com.yong.taximeter.ui.main.subpage.home
+package com.yong.taximeter.ui.main.subscreen.store
 
 import cafe.adriel.voyager.core.model.ScreenModel
 
-class HomeViewModel: ScreenModel {
+class StoreViewModel: ScreenModel {
     // TODO: ViewModel 로직 구현
 }
